refactor: move grid_pool to experimentalist.grid_
