+++ src/autora/experimentalist/pooler/grid.py
@@ -1,14 +1,10 @@
-""""""
 from itertools import product
-from typing import Sequence
+from typing import List
 
-import pandas as pd
-
-from autora.state.delta import Result, wrap_to_use_state
-from autora.variable import Variable, VariableCollection
+from autora.variable import IV
 
 
-def grid_pool_from_ivs(ivs: Sequence[Variable]) -> product:
+def grid_pool(ivs: List[IV]):
     """Creates exhaustive pool from discrete values using a Cartesian product of sets"""
     # Get allowed values for each IV
     l_iv_values = []
@@ -21,90 +17,3 @@ def grid_pool_from_ivs(ivs: Sequence[Variable]) -> product:
 
     # Return Cartesian product of all IV values
     return product(*l_iv_values)
-
-
-def grid_pool_from_variables(variables: VariableCollection) -> pd.DataFrame:
-    """Creates exhaustive pool of conditions given a definition of variables with allowed_values.
-
-    Args:
-        variables: a VariableCollection with `independent_variables` – a sequence of Variable
-        objects, each of which has an attribute `allowed_values` containing a sequence of values.
-
-    Returns: a Result / Delta object with the conditions as a pd.DataFrame in the `conditions` field
-
-    Examples:
-        >>> from autora.state.delta import State
-        >>> from autora.variable import VariableCollection, Variable
-        >>> from dataclasses import dataclass, field
-        >>> import pandas as pd
-        >>> import numpy as np
-
-        With one independent variable "x", and some allowed values, we get exactly those values
-        back when running the executor:
-        >>> grid_pool_from_variables(variables=VariableCollection(
-        ...     independent_variables=[Variable(name="x", allowed_values=[1, 2, 3])]
-        ... ))
-        {'conditions':    x
-        0  1
-        1  2
-        2  3}
-
-        The allowed_values must be specified:
-        >>> grid_pool_from_variables(
-        ...     variables=VariableCollection(independent_variables=[Variable(name="x")]))
-        Traceback (most recent call last):
-        ...
-        AssertionError: gridsearch_pool only supports independent variables with discrete...
-
-        With two independent variables, we get the cartesian product:
-        >>> grid_pool_from_variables(variables=VariableCollection(independent_variables=[
-        ...         Variable(name="x1", allowed_values=[1, 2]),
-        ...         Variable(name="x2", allowed_values=[3, 4]),
-        ... ]))["conditions"]
-           x1  x2
-        0   1   3
-        1   1   4
-        2   2   3
-        3   2   4
-
-        If any of the variables have unspecified allowed_values, we get an error:
-        >>> grid_pool_from_variables(
-        ...     variables=VariableCollection(independent_variables=[
-        ...         Variable(name="x1", allowed_values=[1, 2]),
-        ...         Variable(name="x2"),
-        ... ]))
-        Traceback (most recent call last):
-        ...
-        AssertionError: gridsearch_pool only supports independent variables with discrete...
-
-
-        We can specify arrays of allowed values:
-        >>> grid_pool_from_variables(
-        ...     variables=VariableCollection(independent_variables=[
-        ...         Variable(name="x", allowed_values=np.linspace(-10, 10, 101)),
-        ...         Variable(name="y", allowed_values=[3, 4]),
-        ...         Variable(name="z", allowed_values=np.linspace(20, 30, 11)),
-        ... ]))["conditions"]
-                 x  y     z
-        0    -10.0  3  20.0
-        1    -10.0  3  21.0
-        2    -10.0  3  22.0
-        3    -10.0  3  23.0
-        4    -10.0  3  24.0
-        ...    ... ..   ...
-        2217  10.0  4  26.0
-        2218  10.0  4  27.0
-        2219  10.0  4  28.0
-        2220  10.0  4  29.0
-        2221  10.0  4  30.0
-        <BLANKLINE>
-        [2222 rows x 3 columns]
-
-    """
-    raw_conditions = grid_pool_from_ivs(variables.independent_variables)
-    iv_names = [v.name for v in variables.independent_variables]
-    conditions = pd.DataFrame(raw_conditions, columns=iv_names)
-    return Result(conditions=conditions)
-
-
-grid_pool = wrap_to_use_state(grid_pool_from_variables)
